Move wrapping AbstractRequestHandlerWithTypedInput into AbstractRequestHandler to lower level
