+++ app/domain/environment-agnostic-handlers/all-operations-handler/index.ts
@@ -1,0 +1,2 @@
+import allOperationsHandler from './handler';
+export default allOperationsHandler; // there is no short syntax to re-export default as default
