+++ app/domain/environment-agnostic-handlers/per-operation-handlers/create-payment/index.ts
@@ -1,0 +1,5 @@
+import handler from './handler';
+import { RequestBodySchema, RequestBodySchemaType } from './request-schema';
+import { wrapHandlerToValidateInput } from '../../input-validation';
+
+export default wrapHandlerToValidateInput<RequestBodySchemaType>(handler, RequestBodySchema);
