+++ app/environment-specific-handlers/aws-http/index.ts
@@ -1,9 +1,8 @@
 import { createApiGatewayHandler } from './api-gateway-adapter';
-import {
-  allOperationsHandler as envAgnosticAllOpsHandler
-} from '../../domain/environment-agnostic-handlers/index';
+import { default as envAgnosticAllOpsHandler } from '../../domain/environment-agnostic-handlers/all-operations-handler';
 
 
 export const allOperationsHandler = createApiGatewayHandler(envAgnosticAllOpsHandler);
 
-// Theoretically other (more specific-purpose) handlers can be exported here in case of necessity.
+// Other (lower-level) handlers can be exported here in case of necessity,
+// but that is not needed acccording to the single-function design.
