+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "include": [ "app/**/*.ts" ],
-  "exclude": [ "/**/*.test.ts", "node_modules" ],
+  "exclude": [ "/**/*.spec.ts", "node_modules" ],
   "compilerOptions": {
     "rootDir": ".",
     "target": "es2020",
